tools/c7n_snowflake/pyproject.toml
[project]
authors = [
    {name = "Cloud Custodian Project"},
]
license = {text = "Apache-2.0"}
requires-python = "<4.0.0,>=3.9.2"
dependencies = [
    "c7n",
    "snowflake>=1.7.0",
    "snowflake-connector-python>=3.17.1",
    "snowflake-snowpark-python>=1.37.0",
]
name = "c7n_snowflake"
version = "0.0.1"
description = "Cloud Custodian - Snowflake Support"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Distributed Computing",
]


[project.urls]
homepage = "https://cloudcustodian.io"
repository = "https://github.com/cloud-custodian/cloud-custodian"
documentation = "https://cloudcustodian.io/docs/"

[dependency-groups]
dev = [
    "pytest<8.0",
    "vcrpy>=5.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["c7n_snowflake"]
